# service/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenID Federation Registry API",
    "description": "REST API for managing OpenID Federation entities, policies, modules, and trust marks. This API provides endpoints for creating, reading, updating, and deleting federation entities, policies, trust anchors, resolvers, trust marks, and trust mark subjects.",
    "contact": {
      "name": "Sweden Connect",
      "url": "https://www.swedenconnect.se",
      "email": "support@swedenconnect.se"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.5.10-SNAPSHOT"
  },
  "servers": [
    {
      "url": "https://registry.local.swedenconnect.se",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/registration/v1/{tenant}/{orgNumber}/{registrationId}": {
      "get": {
        "tags": [
          "FederationRegistration"
        ],
        "summary": "Get a single registration by ID",
        "operationId": "getById",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "registrationId",
            "in": "path",
            "description": "Registration ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Registration"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "FederationRegistration"
        ],
        "summary": "Re-run the registration flow for an existing entity",
        "operationId": "updateRegistration",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "registrationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationJoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Registration"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "FederationRegistration"
        ],
        "summary": "Remove a join record",
        "operationId": "deleteJoin",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "registrationId",
            "in": "path",
            "description": "ID of the registration to remove",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/registration/v1/{tenant}/{orgNumber}/{joinId}": {
      "post": {
        "tags": [
          "FederationRegistration"
        ],
        "summary": "Create a registration request on this id",
        "operationId": "createJoinWithId",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "joinId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationJoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Registration"
                }
              }
            }
          }
        }
      }
    },
...
    "/registration/v1/{tenant}/{orgNumber}": {
      "get": {
        "tags": [
          "FederationRegistration"
        ],
        "summary": "List all registration records for current organization",
        "operationId": "listRegistrations",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Registration"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/registration-flow/v1/{tenant}/{orgNumber}/flows": {
      "get": {
        "tags": [
          "Registration Flow"
        ],
        "summary": "List all registration flows",
        "operationId": "listFlows_1",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlowSummaryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/registration-admin/v1/{tenant}/{orgNumber}": {
      "get": {
        "tags": [
          "RegistrationAdmin"
        ],
        "summary": "List all registration records for current organization",
        "operationId": "listRegistrations_1",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Registration"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegistrationJoinRequest": {
        "type": "object",
        "properties": {
          "joinId": {
            "type": "string",
            "format": "uuid",
            "description": "JoinId - Id for the join flow to register on",
            "readOnly": true
          },
          "entityIdentifier": {
            "type": "string",
            "description": "Entity identifier of the applicant",
            "example": "https://example.com/entity"
          },
          "trustmarksRequested": {
            "type": "array",
            "description": "Trustmarks requested",
            "items": {
              "$ref": "#/components/schemas/TrustmarkRequest"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {},
            "description": "FederationMetadata"
          }
        }
      },
      "TrustmarkRequest": {
        "type": "object",
        "properties": {
          "trustmarkIssuer": {
            "type": "string",
            "description": "Entity identifier of the trustmark issuer",
            "example": "https://fed.swedenconnect.se/tmi"
          },
          "trustmarkType": {
            "type": "array",
            "description": "TrustmarkType",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ContextDiffEntry": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Context key that changed"
          },
          "changeType": {
            "type": "string",
            "description": "Type of change: ADDED, CHANGED, or REMOVED"
          },
          "before": {
            "type": "string",
            "description": "Value before execution, null for ADDED entries"
          },
          "after": {
            "type": "string",
            "description": "Value after execution, null for REMOVED entries"
          }
        }
      },
      "Registration": {
        "type": "object",
        "properties": {
          "registrationId": {
            "type": "string",
            "format": "uuid",
            "description": "Id of this specific registration",
            "readOnly": true
          },
          "joinId": {
            "type": "string",
            "format": "uuid",
            "description": "Join ID",
            "readOnly": true
          },
          "entityIdentifier": {
            "type": "string",
            "description": "EntityId for the requested registration",
            "example": "https://example.com/service",
            "readOnly": true
          },
          "intermediateEntityId": {
            "type": "string",
            "description": "EntityId of the Intermediate that the registration is made on",
            "example": "https://example.com/entity",
            "readOnly": true
          },
          "isHosted": {
            "type": "boolean",
            "description": "True if the entity's metadata is hosted in this registry",
            "readOnly": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {},
            "description": "Hosted metadata, present when isHosted is true",
            "readOnly": true
          },
          "registrationType": {
            "type": "string",
            "description": "Registration type: SUBORDINATE or TRUST_MARK_SUBORDINATE",
            "enum": [
              "SUBORDINATE",
              "TRUST_MARK_SUBORDINATE"
            ],
            "readOnly": true
          },
          "statusFedreg": {
            "type": "string",
            "description": "Current federation registration status",
            "enum": [
              "STARTED",
              "PENDING_APPROVAL",
              "APPROVED",
              "REJECTED"
            ],
            "readOnly": true
          },
          "rejectionReason": {
            "type": "string",
            "description": "Reason for rejection, present when status_fedreg is DENY",
            "readOnly": true
          },
          "statusTrustmarks": {
            "type": "array",
            "description": "Trustmark status per requested trustmark",
            "items": {
              "$ref": "#/components/schemas/TrustmarkRegistration"
            },
            "readOnly": true
          },
          "tags": {
            "type": "array",
            "description": "Tags describing the entity type, e.g. OIDC, SAML, RP, OP, IDP, SP",
            "example": [
              "OIDC",
              "RP"
            ],
            "items": {
              "type": "string",
              "enum": [
                "OIDC",
                "SAML",
                "HOSTED",
                "FED",
                "RP",
                "OP",
                "AS",
                "OAC",
                "ORS",
                "SAML_IDP",
                "SAML_SP",
                "SAML_RP"
              ]
            },
            "readOnly": true
          },
          "organizationName": {
            "type": "string",
            "description": "Name of the organization that submitted the registration",
            "readOnly": true
          },
          "subordinateEntityId": {
            "type": "string",
            "description": "Entity ID of the subordinate subject, populated for TRUST_MARK_SUBORDINATE registrations",
            "readOnly": true
          },
          "successful": {
            "type": "boolean",
            "description": "True if the pipeline completed without any step failures"
          },
          "steps": {
            "type": "array",
            "description": "Ordered list of step execution records",
            "items": {
              "$ref": "#/components/schemas/StepExecutionRecord"
            }
          }
        }
      },
      "StepExecutionRecord": {
        "type": "object",
        "properties": {
          "stepName": {
            "type": "string",
            "description": "Display name of the step",
            "example": "LoadEntityConfigurationStep"
          },
          "status": {
            "type": "string",
            "description": "Step execution outcome",
            "example": "SUCCESS"
          },
          "message": {
            "type": "string",
            "description": "Human-readable summary of the step result"
          },
          "issues": {
            "type": "array",
            "description": "Validation issues found during this step",
            "items": {
              "$ref": "#/components/schemas/StepIssue"
            }
          },
          "contextDiff": {
            "type": "array",
            "description": "Changes made to the pipeline context by this step",
            "items": {
              "$ref": "#/components/schemas/ContextDiffEntry"
            }
          }
        }
      },
      "StepIssue": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Field that caused the issue",
            "example": "InternalPreRegistrationStep.status"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the issue"
          },
          "severity": {
            "type": "string",
            "description": "Severity of the issue",
            "example": "ERROR"
          }
        }
      },
      "TrustmarkRegistration": {
        "type": "object",
        "properties": {
          "trustmarkIssuer": {
            "type": "string",
            "description": "Entity identifier of the trustmark issuer",
            "example": "https://fed.swedenconnect.se/tmi"
          },
          "trustmarkStatus": {
            "type": "array",
            "description": "Trustmarks with status",
            "items": {
              "$ref": "#/components/schemas/TrustmarkStatus"
            }
          }
        }
      },
      "TrustmarkStatus": {
        "type": "object",
        "properties": {
          "trustmarkType": {
            "type": "string",
            "description": "Trustmark type identifier (URI)",
            "example": "https://trust.example.com/tm/loa3"
          },
          "status": {
            "type": "string",
            "description": "Current status of this trustmark",
            "enum": [
              "STARTED",
              "PENDING_APPROVAL",
              "APPROVED",
              "REJECTED"
            ]
          }
        }
      },
...
      "FlowSummaryDto": {
        "type": "object",
        "properties": {
          "flowId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "flowType": {
            "type": "string",
            "enum": [
              "INTERMEDIATE",
              "TRUST_MARK_ISSUER"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT token obtained from authentication endpoint",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}